Merge prerelease into master - Dec 23

+++ RMS/DeleteOldObservations.py
@@ -16,11 +16,11 @@ import ephem
 
 from RMS.CaptureDuration import captureDuration
 from RMS.ConfigReader import loadConfigFromDirectory
-from RMS.Logger import initLogging, getLogger
+from RMS.Logger import LoggingManager, getLogger
 from RMS.Misc import RmsDateTime, UTCFromTimestamp
 
 # Get the logger from the main module
-log = getLogger("logger")
+log = getLogger("rmslogger")
 
 
 # Python 2 doesn't have the timestamp function, so make one
@@ -287,7 +287,7 @@ def objectsToDelete(object_path, stationID, quota_gb=0, bz2=False):
         list of full paths to file system objects for deletion
     """
 
-    if quota_gb == 0 or quota_gb == None:
+    if (quota_gb == 0) or (quota_gb is None):
         log.info("Disc quota system disabled for {:s}".format(object_path))
         return []
 
@@ -352,7 +352,8 @@ def rmList(delete_list, dummy_run=True, log_deletions=True):
         full_path = os.path.expanduser(full_path)
         try:
             if dummy_run:
-                log.info("Config setting inhibited deletion of {}".format(os.path.basename(full_path)))
+                if log_deletions:
+                    log.info("Config setting inhibited deletion of {}".format(os.path.basename(full_path)))
             else:
                 if os.path.exists(full_path):
                     if os.path.isdir(full_path):
@@ -730,25 +731,27 @@ def deleteOldObservations(data_dir, captured_dir, archived_dir, config, duration
     log.info('clearing down old data')
     deleteOldDirs(data_dir, config)
 
-    # calculate the captured directory allowance and print to log
-    if (config.rms_data_quota is None or
-        config.arch_dir_quota is None or
-        config.bz2_files_quota is None or
-        config.continuous_capture_quota is None or
-        config.log_files_quota is None):
-        log.info("Deleting files by space quota is not enabled, some quota is None.")
-    else:
-        capt_dir_quota = config.rms_data_quota
-        capt_dir_quota -= config.arch_dir_quota
-        capt_dir_quota -= config.bz2_files_quota
-        capt_dir_quota -= config.continuous_capture_quota
-        capt_dir_quota -= config.log_files_quota
 
-        if capt_dir_quota <= 0:
-            log.warning("No quota allocation remains for captured directories, please increase rms_data_quota")
-            capt_dir_quota = 0
+    if config.quota_management_enabled:
+        # calculate the captured directory allowance and print to log
+        if (config.rms_data_quota is None or
+            config.arch_dir_quota is None or
+            config.bz2_files_quota is None or
+            config.continuous_capture_quota is None or
+            config.log_files_quota is None):
+            log.info("Deleting files by space quota is not enabled, some quota is None.")
+        else:
+            capt_dir_quota = config.rms_data_quota
+            capt_dir_quota -= config.arch_dir_quota
+            capt_dir_quota -= config.bz2_files_quota
+            capt_dir_quota -= config.continuous_capture_quota
+            capt_dir_quota -= config.log_files_quota
 
-        deleteByQuota(archived_dir, capt_dir_quota, captured_dir, config)
+            if capt_dir_quota <= 0:
+                log.warning("No quota allocation remains for captured directories, please increase rms_data_quota")
+                capt_dir_quota = 0
+
+            deleteByQuota(archived_dir, capt_dir_quota, captured_dir, config)
 
     # Calculate the approximate needed disk space for the next night
 
@@ -1154,8 +1157,9 @@ if __name__ == '__main__':
     config = loadConfigFromDirectory(cfg_file, cfg_path)
 
     # Initialize the logger
-    initLogging(config)
-    log = getLogger("logger")
+    log_manager = LoggingManager()
+    log_manager.initLogging(config)
+    log = getLogger("rmslogger")
 
     if not os.path.isdir(config.data_dir):
         log.info('Data Dir not found {}'.format(config.data_dir))
